refact : personality -> personalityQuestion으로 리팩토링
