--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/business/personality/dto/AddKidPersonalityReq.java
@@ -1,4 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.business.personality.dto;
-
-public record AddKidPersonalityReq(int ei, int sn, int tf, int jp, String mbti) {
-}
