--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/business/personality/dto/AddKidPersonalityResp.java
@@ -1,4 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.business.personality.dto;
-
-public record AddKidPersonalityResp(Long kidId, AddKidPersonalityReq req) {
-}
