--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/business/personality/dto/GetQuestionsDetailResp.java
@@ -1,7 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.business.personality.dto;
-
-import com.ureca.idle.idlejpa.question.MBTIType;
-
-public record GetQuestionsDetailResp(Long id, String question, String answer1, String answer2, int value, MBTIType type) {
-
-}
