--- src/main/java/com/ureca/idle/idleapi/idleoriginapi/business/personality/dto/RemoveKidPersonalityResp.java
@@ -1,4 +1,0 @@
-package com.ureca.idle.idleapi.idleoriginapi.business.personality.dto;
-
-public record RemoveKidPersonalityResp() {
-}
