+++ src/main/java/com/ureca/idle/idleapi/idleoriginapi/business/personalityQuestions/PersonalityQuestionManagingUseCase.java
@@ -1,0 +1,7 @@
+package com.ureca.idle.idleapi.idleoriginapi.business.personalityQuestions;
+
+import com.ureca.idle.idleapi.idleoriginapi.business.personalityQuestions.dto.*;
+
+public interface PersonalityQuestionManagingUseCase {
+    GetTestQuestionsResp getPersonalityTestQuestions();
+}
